(feature): fixed the query for rating update

+++ backend/restaurant-backend-app/dnapp/lambdas/restaurant-api/Restaurant.Core/Interfaces/Repositories/ILocationRepository.cs
@@ -7,5 +7,5 @@ public interface ILocationRepository
     Task<IReadOnlyList<Location>> GetLocationsAsync(CancellationToken cancellationToken = default);
     Task<IReadOnlyList<Location>> GetLocationOptionsAsync(CancellationToken cancellationToken = default);
     Task<Location?> GetByIdAsync(string id, CancellationToken cancellationToken = default);
-    Task UpdateUserRatingAsync(string locationId, int newFeedbackRating, CancellationToken ct = default);
+    Task UpdateUserRatingAsync(string locationId, double newFeedbackRating, CancellationToken ct = default);
 }
